Split application helper into 3
