+++ app/controllers/decidim/social_crowdfunding/campaigns_controller.rb
@@ -3,6 +3,9 @@
 module Decidim
   module SocialCrowdfunding
     class CampaignsController < Decidim::SocialCrowdfunding::ApplicationController
+      helper CampaignHelper
+      helper ThermometerHelper
+
       def show; end
     end
   end
